MDEV-19801: Change defaults for CHANGE MASTER TO so that GTID-based replication is used by default if master supports it
This commit makes replicas crash-safe by default by changing the
Using_Gtid value to be Slave_Pos on a fresh slave start and after
RESET SLAVE is issued. If the primary server does not support GTIDs
(i.e., version < 10), the replica will fall back to Using_Gtid=No on
slave start and after RESET SLAVE.

Additional informational messages are added to notify users when
Using_Gtid is automatically changed. That is, if a RESET SLAVE
reverts Using_Gtid back to Slave_Pos, a message is given; and if a
replica falls back to Using_Gtid=No because the primary doesn’t
support GTIDs, an informational note is logged.

This commit only contains the code and infrastructure-related test
changes. Later commits contain the MTR changes which account for
this change.

Reviewed By:
============
Andrei Elkin <andrei.elkin@mariadb.com>