MDEV-21136 InnoDB's records_in_range estimates can be way off
Experimental commit. Get rid of savepoints array. Squash or remove after
testing and code review.

Add the following comment to the previous commit during squashing:

The latching order corresponds to WL#6326 rules, i.e.:

(2.1) [same as (1.1)]: Page latches must be acquired in descending order
of tree level.

(2.2) When acquiring a node pointer page latch at level L, we must hold
the left sibling page latch (at level L) or some ancestor latch
(at level>L).

When we dive to the level down, the parent page is unlatched only after
the the current level page is latched. When we estimate the number of
rows on some level, we latch the left border, then fetch the next page,
and then fetch the next page unlatching the previous page after the
current page is latched until the right border is reached. I.e. the left
sibling is always latched when we acquire page latch on the same level. When we
reach the right border, the current page is unlatched, and then the right
border is latched. We can do this because the right border's parent is latched.